feat: rewrite form arhitecture
