+++ app/src/components/proposals/TransferForm.tsx
@@ -1,0 +1,41 @@
+import React from 'react';
+import { FormGroup, ProposalData } from './CreateProposalPopup';
+
+interface TransferFormProps {
+  proposalForm: ProposalData;
+  handleInputChange: (
+    e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>,
+  ) => void;
+}
+
+export default function TransferForm({
+  proposalForm,
+  handleInputChange,
+}: TransferFormProps) {
+  return (
+    <>
+      <FormGroup>
+        <label htmlFor="receiverId">Receiver ID</label>
+        <input
+          type="text"
+          id="receiverId"
+          name="receiverId"
+          value={proposalForm.receiverId}
+          onChange={handleInputChange}
+          required
+        />
+      </FormGroup>
+      <FormGroup>
+        <label htmlFor="amount">Amount</label>
+        <input
+          type="text"
+          id="amount"
+          name="amount"
+          value={proposalForm.amount}
+          onChange={handleInputChange}
+          required
+        />
+      </FormGroup>
+    </>
+  );
+}
